vector_search/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
sentence-transformers
python-dotenv